% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superwell.R
\name{calc_wells}
\alias{calc_wells}
\title{Need description}
\usage{
calc_wells(time, well_radius, well_params, well_data)
}
\arguments{
\item{time}{Passing through the previously loaded well_parameters data}

\item{well_radius}{Passing through the previously loaded well_data object}

\item{well_params}{initial Q guess}

\item{well_data}{Passing through the main dataframe object}
}
\value{
result - need return info
}
\description{
Need long description
}
