% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superwell.R
\name{calculate_drawdown}
\alias{calculate_drawdown}
\title{Title}
\usage{
calculate_drawdown(well_data, well_parameters, WT, num_iterations, df, i,
  TotTime, t_time)
}
\arguments{
\item{well_data}{fill}

\item{well_parameters}{fill}

\item{WT}{fill}

\item{num_iterations}{fill}

\item{df}{fill}

\item{i}{fill}

\item{TotTime}{fill}

\item{t_time}{fill}
}
\value{
fill
}
\description{
Title
}
