% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superwell.R
\name{calculate_output}
\alias{calculate_output}
\title{Calculate well output data}
\usage{
calculate_output(well_parameters, well_data, rw, df, i, num_iterations)
}
\arguments{
\item{well_parameters}{Passing through the previously loaded well_parameters data}

\item{well_data}{Passing through the previously loaded well_data object}

\item{rw}{initial Q guess}

\item{df}{Passing through the main dataframe object}

\item{i}{Passing through i from the main loop}

\item{num_iterations}{Passing through num_iterations from the main loop}
}
\value{
outputList - returns the formatted list of output variables for writing to excel
}
\description{
This function replaces the while loop located near the end of the main function in previous code
}
\author{
Jason Evanoff; jason.evanoff@pnnl.gov
}
