% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superwell.R
\name{calcWellsTheis}
\alias{calcWellsTheis}
\title{Theis solution}
\usage{
calcWellsTheis(t, rw, wp)
}
\arguments{
\item{t}{annual operational time, comes from wellParams.yml (currently fixed at 365 days i.e., 31536000 s/yr)}

\item{rw}{radius of well. REDFLAG should be radius of influence. CHECK}

\item{wp}{well parameters from wellParams.yml}
}
\value{
drawdown using Theis solution
}
\description{
Calculates drawdown using Theis solution
}
\details{
affected by initial yield, annual operational time, and aquifer properties (T and S)
}
\author{
Superwell team; hassan.niazi@pnnl.gov
}
