% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superwell.R
\name{guess_Q}
\alias{guess_Q}
\title{Guess the radius of influence of Q}
\usage{
guess_Q(num_iterations, well_parameters, well_data, df, i, inRange, WT,
  t_time)
}
\arguments{
\item{num_iterations:}{passed in from previous function - used as a counter}

\item{well_parameters:}{well parameters read in from the well_params file}

\item{well_data:}{well data that is calculated during the run, not read in from the file}

\item{df:}{main input dataframe}

\item{inRange:}{used in the while loop as a tracking mechanism}

\item{WT:}{unsure}
}
\value{
fill
}
\description{
Long description
}
\author{
Jason Evanoff; jason.evanoff@pnnl.gov
}
