% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superwell.R
\name{load_config}
\alias{load_config}
\title{Load electricity data by GCAM basin}
\usage{
load_config(config_file, country = "Iran")
}
\arguments{
\item{config_file}{Full path with file name and extension to the input configuration CSV file}
}
\value{
<fill in>
}
\description{
Read in the cost of electricity by GCAM basin (USD/Unit for 172 countries)
}
\author{
<name>; <email>
}
