% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superwell.R
\name{load_elec_data}
\alias{load_elec_data}
\title{Load electricity data by GCAM basin}
\usage{
load_elec_data(electricity_cost_file)
}
\arguments{
\item{electricity_cost_file}{Full path with file name and extension to the input GCAM electrical rates YAML file}
}
\value{
loaded file
}
\description{
Read in the cost of electricity by GCAM basin (USD/Unit for 172 countries)
}
