% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superwell.R
\name{load_well_data}
\alias{load_well_data}
\title{Load well parameters}
\usage{
load_well_data(well_param_file)
}
\arguments{
\item{well_param_file}{Full path with file name and extension to the input well parameters YAML file}
}
\value{
Data frame of well parameters
}
\description{
Load a total of 14 well-specific parameters (e.g., constants, dimensions, costs, etc.) from a wellparam file
}
\author{
Superwell team; hassan.niazi@pnnl.gov
}
