% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superwell.R
\name{main}
\alias{main}
\title{Main <function to be completely replaced !!!>}
\usage{
main(well_param_file, elec_cost_file, config_file, output_csv)
}
\arguments{
\item{well_param_file}{Full path with file name and extension to the input well parameters YAML file}

\item{elec_cost_file}{Full path with file name and extension to the input GCAM electrical rates YAML file}

\item{config_file}{Full path with file name and extension to the input configuration CSV file}

\item{output_csv}{Full path with file name and extension to the output CSV file}
}
\value{
Data frame of well parameters
}
\description{
Function to be completely replaced
}
\author{
<name>; <email>
}
