% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superwell.R
\name{process_data}
\alias{process_data}
\title{Process the well data input data frame}
\usage{
process_data(df, well_parameters, ec)
}
\arguments{
\item{df}{Passing through the previously loaded data frame}

\item{well_parameters}{Passing through the previously loaded well_parameters object}

\item{ec}{Passing through the main dataframe object}
}
\value{
return_df - returns the row binded dataframe
}
\description{
This function calculates well specific data and delivers the output for writing to csv file
}
\author{
Jason Evanoff; jason.evanoff@pnnl.gov
}
