% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superwell.R
\name{superwell}
\alias{superwell}
\title{Main Superwell function}
\usage{
superwell(well_params, elec_rates, config_file, output_dir)
}
\arguments{
\item{well_params}{load all well parameters from load_well_data function which takes wellParams.yml as an input}

\item{elec_rates}{load all GCAM electricity costs from load_elec_data which takes GCAM_Electrical_Rates.yml as an input}

\item{config_file}{load all aquifer data for a country from load_config function which takes inputs.csv as an input. inputs.csv is generated from digitized maps from 4 datasets on aquifer properties}

\item{output_dir}{specifies output directory}
}
\value{
cost curves of nonrenewable groundwater in basins of a country
}
\description{
Calculates cost of supplying groundwater (i.e. cost curves)
}
\author{
Superwell team; hassan.niazi@pnnl.gov
}
