% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superwell.R
\name{load_config}
\alias{load_config}
\title{Load aquifer data by country}
\usage{
load_config(config_file, country)
}
\arguments{
\item{config_file}{Full path with file name and extension to the input configuration CSV file}
}
\value{
hydro-geological data for each aquifer of the specified country
}
\description{
Read in the aquifer data for countries
}
\author{
Superwell team; hassan.niazi@pnnl.gov
}
