% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superwell.R
\name{load_elec_data}
\alias{load_elec_data}
\title{Load electricity data by GCAM basin}
\usage{
load_elec_data(el_cost_file)
}
\arguments{
\item{el_cost_file}{Full path with file name and extension to the input GCAM electrical rates YAML file}
}
\value{
<fill in>
}
\description{
Read in the cost of electricity by GCAM basin (USD/Unit for 172 countries)
}
\author{
<name>; <email>
}
